creating file structure for code refactoring
