+++ frontend/occupi-mobile4/models/requests.ts
@@ -1,0 +1,1 @@
+//these models describe data to be sent to the api (POST body)
