+++ frontend/occupi-mobile4/models/response.ts
@@ -1,0 +1,1 @@
+//these models describe the structure of the responses from the api
