+++ frontend/occupi-mobile4/utils/auth.ts
@@ -1,0 +1,2 @@
+//this folder contains functions that will call the service functions which make api requests for authentication
+//the purpose of this file is to refine and process the data and return these
